+++ CHANGELOG.md
@@ -7,6 +7,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2021-09-16
+
+### Added
+- Added check for attribute definitions with multiple/duplicate short names.
+
+### Fixed
+- Leading and trailing space characters are now removed from short names.
+
+
 ## [1.1.0] - 2021-09-14
 
 ### Added